+++ core/src/main/java/com/github/glusk2/sprouts/core/comb/SproutsFaces.java
@@ -119,7 +119,7 @@ public final class SproutsFaces {
      * </pre>
      *
      * @param submove A submove in the game of sprouts (see
-     *                {@link com.github.glusk2.sprouts.core.Submove#asEdge()
+     *                {@link com.github.glusk2.sprouts.core.moves.Submove#asEdge()
      *                Submove.asEdge()}).
      *                The submove need not be completed.
      * @return the face in which this {@code submove} is drawn as a set of
