+++ core/src/main/java/com/github/glusk2/sprouts/core/moves/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains classes and interface that model player-drawn
+ * moves and submoves.
+*/
+package com.github.glusk2.sprouts.core.moves;
