+++ core/src/main/java/com/github/glusk2/sprouts/core/snapshots/SproutAdd.java
@@ -4,10 +4,10 @@ import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.math.Vector2;
 import com.github.glusk2.sprouts.core.CobwebSwitch;
-import com.github.glusk2.sprouts.core.Move;
-import com.github.glusk2.sprouts.core.RenderedMove;
 import com.github.glusk2.sprouts.core.comb.SproutsGameState;
 import com.github.glusk2.sprouts.core.comb.SproutsStateAfterMove;
+import com.github.glusk2.sprouts.core.moves.Move;
+import com.github.glusk2.sprouts.core.moves.RenderedMove;
 
 /**
  * This Snapshot represents the game board <em>after</em> a Move is drawn,
